pyproject.toml
[tool.poetry]
name = "audiobook"
version = "0.1.0"
description = ""
authors = ["Robert Clayton <41345902+robert-clayton@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
audiobook = "audiobook.__main__:main"
audiobook-web = "audiobook.web.app:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
transformers = ">=4.33.0,<4.41.0"
accelerate = "^0.34.2"
bitsandbytes = "^0.43.3"
numpy = "<2.0.0"
torch = [
    { url = "https://download.pytorch.org/whl/cu121/torch-2.4.0%2Bcu121-cp311-cp311-linux_x86_64.whl", markers = "sys_platform == 'linux'" },
    { url = "https://download.pytorch.org/whl/cu121/torch-2.4.0%2Bcu121-cp311-cp311-win_amd64.whl", markers = "sys_platform == 'win32'" }
]
torchaudio = [
    { url = "https://download.pytorch.org/whl/cu121/torchaudio-2.4.0%2Bcu121-cp311-cp311-linux_x86_64.whl", markers = "sys_platform == 'linux'" },
    { url = "https://download.pytorch.org/whl/cu121/torchaudio-2.4.0%2Bcu121-cp311-cp311-win_amd64.whl", markers = "sys_platform == 'win32'" }
]
torchvision = [
    { url = "https://download.pytorch.org/whl/cu121/torchvision-0.19.0%2Bcu121-cp311-cp311-linux_x86_64.whl", markers = "sys_platform == 'linux'" },
    { url = "https://download.pytorch.org/whl/cu121/torchvision-0.19.0%2Bcu121-cp311-cp311-win_amd64.whl", markers = "sys_platform == 'win32'" }
]
sentencepiece = "^0.2.0"
protobuf = "^5.28.1"
huggingface-hub = "^0.24.7"
hqq = "^0.2.2"
bs4 = "^0.0.2"
requests = "^2.32.3"
coqui-tts = "^0.24.1"
nltk = "^3.9.1"
tqdm = "^4.66.5"
python-dotenv = "^1.0.1"
diffusers = {git = "https://github.com/huggingface/diffusers.git"}
cloudscraper = "^1.2.71"
flask = "^3.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
psutil = "^5.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=audiobook",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["audiobook"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchaudio.*",
    "torchvision.*",
    "transformers.*",
    "diffusers.*",
    "coqui_tts.*",
    "nltk.*",
    "bs4.*",
    "cloudscraper.*"
]
ignore_missing_imports = true
